backend of bloglist

+++ controllers/blogs.js
@@ -1,39 +1,72 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs)
+  const blogs = await Blog
+  .find({}).populate('user', { username: 1, name: 1})
+  response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
 
-    try {
-        const result = await blog.save()
-        response.status(201).json(result)
-    }
-    catch (error) { response.status(400).send(error.message) }
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const token = request.token
+  const userToken = request.user
+  console.log(userToken.id)
+  if (!token || !userToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid '})
+  }
+
+  const user = await User.findById(userToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id,
+  })
+
+ // try {
+    const savedBlog = await blog.save()
+    
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
+    response.json(savedBlog)
+  //}
+ // catch (error) { response.status(400).send(error.message) }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const token = request.token
+  const userToken = request.user
+  if (blog.user.toString() === userToken.id.toString()) {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+  }
+  else {
+    response.status(401).json({ error: 'invalid token'})
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-    const body = request.body
+  const body = request.body
 
-    const blog = {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes
-    }
+  const blog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes
+  }
 
-    const updatedNote = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  const updatedNote = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
 
-    response.json(updatedNote)
+  response.json(updatedNote)
 })
 
 module.exports = blogsRouter
